No longer require having the no of sweeps in the filename for fitness. They may differ. Also still trying to work out how the heck you input data for fitness

+++ lib/tlearn.rb
@@ -18,12 +18,18 @@ class TLearn
   end
   
   def fitness(data, number_of_sweeps = DEFAULT_NUMBER_OF_SWEEPS)
+    raise "Train me first!" unless network_trained?
+      
     FileUtils.rm_f("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.output")
-    
+
+    setup_fitness_data({data => []})
+        
     execute_tlearn_fitness(number_of_sweeps)
     
     output = File.read("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.output")
     output = output.split("\n").map{|line| line.split("\t").map{|number| number.strip}}
+    output.each{|o| p o}
+    p
     output[-1]
   end
   
@@ -31,12 +37,13 @@ class TLearn
     clear_previous_session
       
     setup_config(training_data)
-    setup_data(training_data)
+    setup_training_data(training_data)
 
     execute_tlearn_train(number_of_sweeps)
-      
+          
     if training_successful?(number_of_sweeps)
       puts File.read("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.#{number_of_sweeps}.wts").split("\n")
+      `cp #{WORKING_DIR}/#{TLEARN_NAMESPACE}.#{number_of_sweeps}.wts #{WORKING_DIR}/#{TLEARN_NAMESPACE}.wts`
     else
       puts("[Error] Training failed") 
     end
@@ -51,9 +58,13 @@ class TLearn
   def training_successful?(number_of_sweeps)
     File.exists?("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.#{number_of_sweeps}.wts")
   end
+
+  def network_trained?
+    File.exists?("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.wts")
+  end
   
   def execute_tlearn_fitness(number_of_sweeps)
-  `cd #{WORKING_DIR} && #{TLEARN_EXECUTABLE} -f #{TLEARN_NAMESPACE} -l evaluator.#{number_of_sweeps}.wts -s #{number_of_sweeps} #{VERIFY_OUTPUTS_ON_EACH_SWEEP} > evaluator.output`
+    `cd #{WORKING_DIR} && #{TLEARN_EXECUTABLE} -f #{TLEARN_NAMESPACE} -l evaluator.wts -s #{number_of_sweeps} #{VERIFY_OUTPUTS_ON_EACH_SWEEP} > evaluator.output`
   end
 
   def execute_tlearn_train(number_of_sweeps)
@@ -64,7 +75,11 @@ class TLearn
     File.open("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.cf",    "w") {|f| f.write(evaulator_config(training_data))}
   end
 
-  def setup_data(training_data)
+  def setup_fitness_data(data)
+    File.open("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.data",  "w") {|f| f.write(build_data(data))}
+  end
+
+  def setup_training_data(training_data)
     #File.open("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.reset", "w") {|f| f.write(reset_config)}
     File.open("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.data",  "w") {|f| f.write(build_data(training_data))}
     File.open("#{WORKING_DIR}/#{TLEARN_NAMESPACE}.teach", "w") {|f| f.write(build_teach_data(training_data))}
